test: 학습테스트 String 완료
